+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add filter for the instance source. Refs UIIN-1132.
 * Increment `@folio/plugin-find-instance` to `v3.0` for `@folio/stripes` `v4` compatibility.
 * Clear detail record pane after new search is performed. Refs UIIN-1074.
+* Add effective location to item list table. Refs UIIN-1124.
 
 ## [3.0.2](https://github.com/folio-org/ui-inventory/tree/v3.0.2) (2020-06-23)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v3.0.1...v3.0.2)